+++ src/main/transfer/Resources/modules/prop-types.js
@@ -1,40 +1,50 @@
 import {PropTypes as T} from 'prop-types'
 
-const ImportFile = {
+const Property = {
   propTypes: {
-    id: T.string.isRequired,
-    status: T.string.isRequired,
-    action: T.string,
-    meta: T.shape({
-      createdAt: T.string,
-      creator: T.object // a User object
-    }),
-    executionDate: T.string,
-    file: T.object // a PublicFile object
+    name: T.string.isRequired,
+    type: T.string,
+    description: T.string,
+    required: T.bool,
+    isArray: T.bool
   },
   defaultProps: {
-
+    required: false,
+    isArray: false
   }
 }
 
-const ExportFile = {
+const OneOf = {
   propTypes: {
-    id: T.string.isRequired,
-    status: T.string.isRequired,
-    action: T.string,
-    meta: T.shape({
-      createdAt: T.string,
-      creator: T.object // a User object
-    }),
-    executionDate: T.string
+    oneOf: T.arrayOf(T.shape(
+      Property.propTypes
+    )).isRequired,
+    description: T.string, // it should not be used because it always contains the same auto generated string
+    required: T.bool
   },
   defaultProps: {
-
+    oneOf: [],
+    required: false
   }
 }
+
+const Schema = {
+  propTypes: {
+    properties: T.arrayOf(T.oneOfType([
+      T.shape(Property.propTypes),
+      T.shape(OneOf.propTypes)
+    ])),
+    identifiers: T.array
+  },
+  defaultProps: {
+    properties: [],
+    identifiers: []
+  }
+}
+
+
 
 
 export {
-  ImportFile,
-  ExportFile
+  Schema
 }
